draftcodebt1stprompt/requirements.txt
# Requirements for Agile Education Analysis Framework
# Python 3.8+ required

# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0
webvtt-py>=0.4.6
python-dateutil>=2.8.2

# Ukrainian NLP Support
spacy>=3.4.0
stanza>=1.4.0
pymorphy3>=1.2.0
pymorphy3-dicts-uk>=2.4.1
ukrainian-word-stress>=1.0.0

# Transformer Models for Ukrainian
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Text Analysis
scikit-learn>=1.3.0
gensim>=4.3.0
nltk>=3.8
textblob>=0.17.0

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.0
networkx>=3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
jsonlines>=3.1.0

# Optional: For advanced speaker diarization
# pyannote.audio>=2.1.0
# resemblyzer>=0.1.1

# Optional: For PDF report generation
# reportlab>=4.0.0
# fpdf2>=2.7.0
